Changes to test RESTful state transfers
Added NewSalesQuote model and controller and used the dependency
injection to set a query so that when a get is done on NewSalesQuote
only those with a status of 'N' are retrieved.  This technique could
also be explored for tenantID, ACL (access control list) entry level
security.  Would like to be able to do this only with a new controller
and using the same POCO